# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stayinn API",
    "version": "1.0.0",
    "description": "API Specification for Stayinn — Hotel Booking Marketplace"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://stayinn.ng/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/upload": {
      "post": {
        "summary": "Upload a file (e.g. image)",
        "tags": ["Uploads"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "File uploaded successfully" }
        }
      }
    },
...
  }
}